requirements.txt
requests
python-dotenv
pymysql
SQLAlchemy
flask